pyproject.toml
[project]
name = "azure-rag-app"
version = "0.1.0"
description = "Retrieval-Augmented Generation app with Azure OpenAI and Azure AI Search using FastAPI"
authors = [{ name = "AI Assistant" }]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.1",
  "azure-ai-search-documents>=11.4.0",
  "azure-identity>=1.17.1",
  "openai>=1.35.3",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.2.2"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
